Actualizar tipo de datos de 'creditos' a número en asignatura y ajustar cálculo en fhistorial

+++ packages/files/src/adapters/php/formatos/fhistorial.php
@@ -251,7 +251,7 @@ foreach ($calificacionCiclo as $ciclos => $ciclo) {
     }
 
     if (is_numeric($detalle["calificacion"]) && $detalle["calificacion"] >= $programa["calificacionAprobatoria"]) {
-      $total_creditos += (int) ($detalle["asignatura"]["creditos"] ?? 0);
+      $total_creditos += (float) ($detalle["asignatura"]["creditos"] ?? 0);
       $total_calificaciones += (float) ($detalle["calificacion"] ?? 0);
       $total_materias += 1;
     }
